@@ -1,6 +1,6 @@
 ////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2016 Realm Inc.
+// Copyright 2019 Realm Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,13 +16,13 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-const { expect } = require("chai");
+import { expect } from "chai";
 
-const PersonAndDogsSchema = require("./schemas/person-and-dogs");
+import { PersonAndDogSchema } from "./schemas/person-and-dogs";
 
-describe("Realm._updateSchema", () => {
+describe("realm._updateSchema", () => {
     it("is a function", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         expect(realm.schema).to.be.an("array");
         // There is a function defined on the Realm
         expect(realm._updateSchema).to.be.a("function");
@@ -33,11 +33,11 @@ describe("Realm._updateSchema", () => {
     });
 
     it("creates a class schema from a name", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
-                { name: "MyClass", properties: { } },
+                { name: "MyClass", properties: {} }
             ]);
         });
         const classNames = realm.schema.map(s => s.name);
@@ -45,11 +45,11 @@ describe("Realm._updateSchema", () => {
     });
 
     it("creates a class schema from a name and properties", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
-                { name: "MyClass", properties: { myField: "string" } },
+                { name: "MyClass", properties: { myField: "string" } }
             ]);
         });
         const MyClassSchema = realm.schema.find(s => s.name === "MyClass");
@@ -57,31 +57,78 @@ describe("Realm._updateSchema", () => {
             name: "MyClass",
             properties: {
                 myField: {
-                    type: "string",
                     indexed: false,
                     name: "myField",
                     optional: false,
+                    type: "string"
+                }
+            }
+        });
+    });
+
+    it("creates a property on an existing class", () => {
+        const realm = new Realm({ schema: PersonAndDogSchema });
+        // Copy the schema
+        const updatedSchema = [...realm.schema];
+        // Locate the Dog schema
+        const dogSchema = updatedSchema.find(s => s.name === "Dog");
+        // Add a fields property
+        dogSchema.properties.friends = "Dog[]";
+        // Update the schema
+        realm.write(() => {
+            realm._updateSchema(updatedSchema);
+        });
+
+        const ModifiedDogSchema = realm.schema.find(s => s.name === "Dog");
+        expect(ModifiedDogSchema).to.deep.equal({
+            name: "Dog",
+            properties: {
+                age: {
+                    indexed: false,
+                    name: "age",
+                    optional: false,
+                    type: "int"
                 },
-            },
+                friends: {
+                    indexed: false,
+                    name: "friends",
+                    objectType: "Dog",
+                    optional: false,
+                    type: "list"
+                },
+                name: {
+                    indexed: false,
+                    name: "name",
+                    optional: false,
+                    type: "string"
+                },
+                owner: {
+                    indexed: false,
+                    name: "owner",
+                    objectType: "Person",
+                    optional: true,
+                    type: "object"
+                }
+            }
         });
     });
 
     it("can use a newly added class", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
-                { name: "MyClass", properties: { myField: "string" } },
+                { name: "MyClass", properties: { myField: "string" } }
             ]);
             realm.create("MyClass", { myField: "some string" });
-            const myClassObjects = realm.objects("MyClass");
+            const myClassObjects = realm.objects<any>("MyClass");
             expect(myClassObjects).to.be.of.length(1);
             expect(myClassObjects[0].myField).to.equal("some string");
         });
     });
 
-    it("fires the schema change event", (done) => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+    it("fires the schema change event", done => {
+        const realm = new Realm({ schema: PersonAndDogSchema });
         realm.addListener("schema", () => {
             expect(realm.schema).to.have.length(3);
             const objectSchemaNames = realm.schema.map(s => s.name);
@@ -92,47 +139,47 @@ describe("Realm._updateSchema", () => {
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
-                { name: "MyClass", properties: { myField: "string" } },
+                { name: "MyClass", properties: { myField: "string" } }
             ]);
         });
     });
 
     it("throws if creating a class schema outside of a transaction", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         expect(() => {
             realm._updateSchema([
                 ...realm.schema,
-                { name: "MyClass", properties: {} },
+                { name: "MyClass", properties: {} }
             ]);
         }).to.throw("Can only create object schema within a transaction.");
     });
 
     it("throws if asked to create a class that already exists", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         expect(() => {
             realm.write(() => {
                 realm._updateSchema([
                     ...realm.schema,
-                    { name: "Person", properties: {} },
+                    { name: "Person", properties: {} }
                 ]);
             });
         }).to.throw("Type 'Person' appears more than once in the schema.");
     });
 
     it("throws if called without a schema object", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         expect(() => {
             realm.write(() => {
-                realm._updateSchema();
+                (realm as any)._updateSchema();
             });
         }).to.throw("Invalid arguments: 1 expected, but 0 supplied.");
     });
 
     it("throws if called with an unexpected type", () => {
-        const realm = new Realm({ schema: PersonAndDogsSchema });
+        const realm = new Realm({ schema: PersonAndDogSchema });
         expect(() => {
             realm.write(() => {
-                realm._updateSchema("w00t");
+                (realm as any)._updateSchema("w00t");
             });
         }).to.throw("schema must be of type 'array', got (w00t)");
     });
