@@ -90,6 +90,9 @@ function fetchWithTimeout(input, init) {
                 url: xhr.responseURL,
                 headers: xhr.responseHeaders
             };
+            if (!options.headers) {
+                options.headers = {'content-type': xhr.getResponseHeader('content-type')};
+            }
             const body = 'response' in xhr ? xhr.response : xhr.responseText;
             resolve(new Response(body, options));
         };
@@ -97,7 +100,7 @@ function fetchWithTimeout(input, init) {
         xhr.ontimeout = () => reject(new TypeError('Network request failed'));
         xhr.open(request.method, request.url, true);
         request.headers.forEach((value, name) => xhr.setRequestHeader(name, value));
-        xhr.send(typeof request._bodyInit === 'undefined' ? null : request._bodyInit);
+        xhr.send(typeof request._bodyInit === 'undefined' ? init.body : request._bodyInit);
     });
 }
 
@@ -331,34 +334,34 @@ function _authenticate(userConstructor, server, json, retries) {
         timeout: 5000
     };
 
-    let request = performFetch(url, options)
-    if (retries < 3) {
-        // Retry on network errors (which are different from the auth endpoint returning an error)
-        request = request.catch(() => {
+    return performFetch(url, options).then((response) => {
+        const contentType = response.headers.get('Content-Type');
+        if (contentType.indexOf('application/json') === -1) {
+            return response.text().then((body) => {
+                throw new AuthError({
+                    title: `Could not authenticate: Realm Object Server didn't respond with valid JSON`,
+                    body,
+                });
+            });
+        } else if (!response.ok) {
+            return response.json().then((body) => Promise.reject(new AuthError(body)));
+        } else {
+            return response.json().then((body) => {
+                // TODO: validate JSON
+                const token = body.refresh_token.token;
+                const identity = body.refresh_token.token_data.identity;
+                const isAdmin = body.refresh_token.token_data.is_admin;
+                return userConstructor.createUser(server, identity, token, false, isAdmin);
+            });
+        }
+    }, (err) => {
+        if (retries < 3) {
+            // Retry on network errors (which are different from the auth endpoint returning an error)
             return _authenticate(userConstructor, server, json, retries + 1);
-        });
-    }
-    return request.then((response) => {
-            const contentType = response.headers.get('Content-Type');
-            if (contentType.indexOf('application/json') === -1) {
-                return response.text().then((body) => {
-                    throw new AuthError({
-                        title: `Could not authenticate: Realm Object Server didn't respond with valid JSON`,
-                        body,
-                    });
-                });
-            } else if (!response.ok) {
-                return response.json().then((body) => Promise.reject(new AuthError(body)));
-            } else {
-                return response.json().then(function (body) {
-                    // TODO: validate JSON
-                    const token = body.refresh_token.token;
-                    const identity = body.refresh_token.token_data.identity;
-                    const isAdmin = body.refresh_token.token_data.is_admin;
-                    return userConstructor.createUser(server, identity, token, false, isAdmin);
-                });
-            }
-        });
+        } else {
+            throw err;
+        }
+    });
 }
 
 function _updateAccount(userConstructor, server, json) {
